Cleanups and refactorings
