+++ Source/Anthropic.Net/GlobalSuppressions.cs
@@ -25,3 +25,12 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("StyleCop.CSharp.NamingRules", "SA1310:Field names should not contain underscore", Justification = "Special & rare exception for version-named variables.", Scope = "member", Target = "~F:Anthropic.Net.Constants.AnthropicModels.ClaudeInstant_v1")]
 [assembly: SuppressMessage("Naming", "CA1707:Identifiers should not contain underscores", Justification = "Special & rare exception for version-named variables.", Scope = "member", Target = "~F:Anthropic.Net.Constants.AnthropicModels.ClaudeInstant_v1_0")]
 [assembly: SuppressMessage("StyleCop.CSharp.NamingRules", "SA1310:Field names should not contain underscore", Justification = "Special & rare exception for version-named variables.", Scope = "member", Target = "~F:Anthropic.Net.Constants.AnthropicModels.ClaudeInstant_v1_0")]
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.CompletionRequest.#ctor(System.String,System.String,System.Int32)")]
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.Models.Messages.ImageSource.#ctor(System.String,System.String,System.String)")]
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.Models.Messages.Message.#ctor(System.String,System.Object)")]
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.Models.Messages.MessageRequest.#ctor(System.String,System.Collections.Generic.IList{Anthropic.Net.Models.Messages.Message},System.Int32)")]
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.Models.Messages.ResponseFormat.#ctor(System.String)")]
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.Models.Messages.Tool.#ctor(System.String,System.String,System.Object)")]
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.ValidationException.#ctor(System.String)")]
+[assembly: SuppressMessage("Style", "IDE0021:Use expression body for constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.Models.Messages.ResponseFormat.#ctor(System.String)")]
+[assembly: SuppressMessage("Style", "IDE0021:Use expression body for constructor", Justification = "<Pending>", Scope = "member", Target = "~M:Anthropic.Net.Models.Messages.ToolUseContentBlock.#ctor")]
