+++ Source/Anthropic.Net/Models/Messages/MessageRequest.cs
