+++ Source/Anthropic.Net/Models/Messages/Streaming/StreamingEvents/MessageStopEvent.cs
@@ -1,0 +1,7 @@
+namespace Anthropic.Net.Models.Messages.Streaming.StreamingEvents;
+
+using Anthropic.Net.Models.Messages.Streaming;
+
+public class MessageStopEvent : MessageStreamEvent
+{
+}
