+++ Source/Anthropic.Net/Models/Messages/Streaming/StreamingEvents/PingEvent.cs
@@ -1,0 +1,10 @@
+namespace Anthropic.Net.Models.Messages.Streaming.StreamingEvents;
+
+using Anthropic.Net.Models.Messages.Streaming;
+
+/// <summary>
+/// Represents a ping event in the message streaming process.
+/// </summary>
+public class PingEvent : MessageStreamEvent
+{
+}
