+++ Source/Anthropic.Net/Models/Messages/Tool.cs
@@ -38,45 +38,3 @@ public class Tool
     [JsonPropertyName("input_schema")]
     public object InputSchema { get; set; }
 }
-
-/// <summary>
-/// Represents a tool call in a message.
-/// </summary>
-public class ToolCall
-{
-    /// <summary>
-    /// Gets or sets the ID of the tool call.
-    /// </summary>
-    [JsonPropertyName("id")]
-    public string Id { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Gets or sets the name of the tool.
-    /// </summary>
-    [JsonPropertyName("name")]
-    public string Name { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Gets or sets the input to the tool.
-    /// </summary>
-    [JsonPropertyName("input")]
-    public object Input { get; set; } = new();
-}
-
-/// <summary>
-/// Represents a tool use in a message.
-/// </summary>
-public class ToolUse
-{
-    /// <summary>
-    /// Gets or sets the type of the content block.
-    /// </summary>
-    [JsonPropertyName("type")]
-    public string Type { get; set; } = "tool_use";
-
-    /// <summary>
-    /// Gets or sets the tool call.
-    /// </summary>
-    [JsonPropertyName("tool_use")]
-    public ToolCall ToolCall { get; set; } = new ToolCall();
-}
