+++ Tests/Anthropic.Net.Tests/Anthropic.Net.Test.csproj
@@ -7,11 +7,11 @@
 
   <ItemGroup>
     <PackageReference Include="NSubstitute" Version="5.3.0" />
-    <PackageReference Include="coverlet.collector" Version="6.0.2" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
+    <!--<PackageReference Include="coverlet.collector" Version="6.0.2" PrivateAssets="all" />-->
+    <!--<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />-->
     <PackageReference Include="Shouldly" Version="4.2.1" />
     <PackageReference Include="xunit" Version="2.9.2" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="3.0.0" PrivateAssets="all" />
+    <!--<PackageReference Include="xunit.runner.visualstudio" Version="3.0.0" PrivateAssets="all" />-->
   </ItemGroup>
